+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 - **fix:** adjust specification concerning added functionality
 - **doc:** adjust usage chapter concerning added functionality
 - **fix:** disable multiple registration
+- **fix:** rename options.dir into options.routes